chore: rename packages and files
