+++ adapters/output/cache/cache_output.go
@@ -1,4 +1,4 @@
-package redis
+package cache
 
 import (
 	"errors"
