+++ adapters/output/db/user_output.go
@@ -1,4 +1,4 @@
-package postgres
+package db
 
 import (
 	"fmt"
